+++ azure/fedora-coreos/kubernetes/lb.tf
@@ -18,7 +18,7 @@ resource "azurerm_dns_aaaa_record" "apiserver" {
   # DNS record
   name = var.cluster_name
   ttl  = 300
-  # IPv4 address of apiserver load balancer
+  # IPv6 address of apiserver load balancer
   records = [azurerm_public_ip.frontend-ipv6.ip_address]
 }
 
@@ -74,6 +74,8 @@ resource "azurerm_lb_rule" "apiserver-ipv4" {
 }
 
 resource "azurerm_lb_rule" "apiserver-ipv6" {
+  count = var.enable_ipv6_load_balancing ? 1 : 0
+
   name                           = "apiserver-ipv6"
   loadbalancer_id                = azurerm_lb.cluster.id
   frontend_ip_configuration_name = "frontend-ipv6"
@@ -113,6 +115,8 @@ resource "azurerm_lb_rule" "ingress-https-ipv4" {
 }
 
 resource "azurerm_lb_rule" "ingress-http-ipv6" {
+  count = var.enable_ipv6_load_balancing ? 1 : 0
+
   name                           = "ingress-http-ipv6"
   loadbalancer_id                = azurerm_lb.cluster.id
   frontend_ip_configuration_name = "frontend-ipv6"
@@ -126,6 +130,8 @@ resource "azurerm_lb_rule" "ingress-http-ipv6" {
 }
 
 resource "azurerm_lb_rule" "ingress-https-ipv6" {
+  count = var.enable_ipv6_load_balancing ? 1 : 0
+
   name                           = "ingress-https-ipv6"
   loadbalancer_id                = azurerm_lb.cluster.id
   frontend_ip_configuration_name = "frontend-ipv6"
@@ -140,7 +146,7 @@ resource "azurerm_lb_rule" "ingress-https-ipv6" {
 
 # Backend Address Pools
 
-# Address pool of controllers
+# Address pools for controllers
 resource "azurerm_lb_backend_address_pool" "controller-ipv4" {
   name            = "controller-ipv4"
   loadbalancer_id = azurerm_lb.cluster.id
@@ -151,7 +157,7 @@ resource "azurerm_lb_backend_address_pool" "controller-ipv6" {
   loadbalancer_id = azurerm_lb.cluster.id
 }
 
-# Address pool of workers
+# Address pools for workers
 resource "azurerm_lb_backend_address_pool" "worker-ipv4" {
   name            = "worker-ipv4"
   loadbalancer_id = azurerm_lb.cluster.id